+++ 1.21.9/src/main/java/com/bleudev/ppl_utils/mixin/client/InventoryScreenMixin.java
@@ -1,0 +1,17 @@
+package com.bleudev.ppl_utils.mixin.client;
+
+import com.bleudev.ppl_utils.util.helper.DiamondHelper;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gui.screen.ingame.InventoryScreen;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(InventoryScreen.class)
+public class InventoryScreenMixin {
+    @Inject(method = "render", at = @At("TAIL"))
+    private void renderCounter(DrawContext context, int mouseX, int mouseY, float deltaTicks, CallbackInfo ci) {
+        DiamondHelper.renderCounter(context, false);
+    }
+}