+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version updates
-[/] \[DEV] Update to Gradle 9.2.1
+[/] \[DEV] Update to Gradle 9.2.1\
+[/] \[DEV] Set Fabric Loom version to 1.14.10
 ## New features
 [+] Restart bar appearing sound\
 [/] \[DEV] Client thinks that it is on Pepeland in development environment