+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## New features
-[+] Ping indicator ([#39](https://github.com/bleudev/ppl_utils/pull/39) bleudev)\
-[+] Show ping key ([#39](https://github.com/bleudev/ppl_utils/pull/39) bleudev)
+[+] Diamond counter (bleudev)