+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## New features
 [+] Diamond counter (bleudev)
+
+## Bug fixes
+[/] \[1.21.6-1.21.8 only] Pepeland Utils Data (`ppl_utils_data.json`) resets every startup